+++ src/condor/types/strings/string.h
@@ -33,7 +33,7 @@ namespace internal{
 		void SetPath(const char* path){filePath = path;}
 		const char* GetPath(){return filePath.c_str();}
 		std::string GetPathStr(){return filePath;}
-		int Length(){return value.length();}
+		int Length(){return (int) value.length();}
 		bool IsEmpty(){return value.empty() || Length() == 0;}
 		static bool Replace(std::string& str, const std::string& from, const std::string& to);
 		static int NthSubStr(int n, const std::string& s, const std::string& p);
